Fix build issues with RevenueService
- Fixed type conversion issues in RevenueService
- Changed appConfig.getCpmRate() and getRevenueShare() to BigDecimal.valueOf()
- Added RevenueService mock to test classes
- Fixed import statements in test files
- Project now builds successfully

+++ src/main/java/com/linksplit/service/RevenueService.java
@@ -32,12 +32,12 @@ public class RevenueService {
         // Get CPM rate (custom or default)
         BigDecimal cpmRate = user.getCustomCpmRate() != null 
                 ? user.getCustomCpmRate() 
-                : appConfig.getCpmRate();
+                : BigDecimal.valueOf(appConfig.getCpmRate());
         
         // Get revenue share (custom or default)
         BigDecimal revenueShare = user.getCustomRevenueShare() != null 
                 ? user.getCustomRevenueShare() 
-                : appConfig.getRevenueShare();
+                : BigDecimal.valueOf(appConfig.getRevenueShare());
         
         // Calculate: (views / 1000) * cpmRate * revenueShare
         return BigDecimal.valueOf(viewCount)
@@ -82,7 +82,7 @@ public class RevenueService {
     public BigDecimal getEffectiveCpmRate(User user) {
         return user.getCustomCpmRate() != null 
                 ? user.getCustomCpmRate() 
-                : appConfig.getCpmRate();
+                : BigDecimal.valueOf(appConfig.getCpmRate());
     }
     
     /**
@@ -91,6 +91,6 @@ public class RevenueService {
     public BigDecimal getEffectiveRevenueShare(User user) {
         return user.getCustomRevenueShare() != null 
                 ? user.getCustomRevenueShare() 
-                : appConfig.getRevenueShare();
+                : BigDecimal.valueOf(appConfig.getRevenueShare());
     }
 }
